Change RPM frame

+++ Src/Libs/_database.c
@@ -7,7 +7,7 @@
 
 #include "_database.h"
 
-uint16_t DB_MCU_RPM = 0;
+uint32_t DB_MCU_RPM = 0;
 uint8_t DB_MCU_Temperature = 0;
 uint8_t DB_BMS_SoC = 100;
 uint8_t DB_BMS_Temperature = 0;
